+++ tests/Mqtt.Client.UnitTests/MqttSubscriptionOverflowTests.cs
@@ -10,10 +10,10 @@ public class MqttSubscriptionOverflowTests
         var sub = TestSubscription.Create(
             "a",
             new MqttSubscriptionOptions { Capacity = 2, Overflow = MqttOverflowMode.DropOldest });
-        sub.Writer.TryWrite(new MqttMessage { Topic = "a", PayloadMemory = new byte[] { 1 } });
-        sub.Writer.TryWrite(new MqttMessage { Topic = "a", PayloadMemory = new byte[] { 2 } });
-        sub.Writer.TryWrite(new MqttMessage { Topic = "a", PayloadMemory = new byte[] { 3 } });
-        sub.Writer.TryComplete();
+        sub.Writer!.TryWrite(new MqttMessage { Topic = "a", PayloadMemory = new byte[] { 1 } });
+        sub.Writer!.TryWrite(new MqttMessage { Topic = "a", PayloadMemory = new byte[] { 2 } });
+        sub.Writer!.TryWrite(new MqttMessage { Topic = "a", PayloadMemory = new byte[] { 3 } });
+        sub.Writer!.TryComplete();
 
         var seen = new List<byte>();
         await foreach (var m in sub.Reader.ReadAllAsync())
@@ -34,10 +34,10 @@ public class MqttSubscriptionOverflowTests
         var sub = TestSubscription.Create(
             "a",
             new MqttSubscriptionOptions { Capacity = 2, Overflow = MqttOverflowMode.DropNewest });
-        sub.Writer.TryWrite(new MqttMessage { Topic = "a", PayloadMemory = new byte[] { 1 } });
-        sub.Writer.TryWrite(new MqttMessage { Topic = "a", PayloadMemory = new byte[] { 2 } });
-        sub.Writer.TryWrite(new MqttMessage { Topic = "a", PayloadMemory = new byte[] { 3 } });
-        sub.Writer.TryComplete();
+        sub.Writer!.TryWrite(new MqttMessage { Topic = "a", PayloadMemory = new byte[] { 1 } });
+        sub.Writer!.TryWrite(new MqttMessage { Topic = "a", PayloadMemory = new byte[] { 2 } });
+        sub.Writer!.TryWrite(new MqttMessage { Topic = "a", PayloadMemory = new byte[] { 3 } });
+        sub.Writer!.TryComplete();
 
         var seen = new List<byte>();
         await foreach (var m in sub.Reader.ReadAllAsync())